Split `mod util` into sub-modules
